clean up file structure and fix gesture handler
