+++ frontEnd/hungr/package.json
@@ -13,8 +13,8 @@
     "react": "~16.11.0",
     "react-dom": "~16.11.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-38.0.2.tar.gz",
-    "react-native-web": "~0.11.7",
-    "react-native-gesture-handler": "1.6.0"
+    "react-native-gesture-handler": "~1.6.0",
+    "react-native-web": "~0.11.7"
   },
   "devDependencies": {
     "@babel/core": "^7.8.6",
