+++ frontEnd/hungr/src/components/RestaurantView.js
@@ -1,12 +1,13 @@
-import React, { Component } from 'react';
-import { View, Text, StyleSheet, Image } from 'react-native';
+import React from 'react';
+import { StyleSheet, Text, View, Image} from 'react-native';
 import { isRequired } from 'react-native/Libraries/DeprecatedPropTypes/DeprecatedColorPropType';
 
-class Card extends Component {
-  render(){
-    return(
-    <View style={styles.Card}>
-      <Image source = {require("./assets/chipotle.jpg")} style={styles.Image}/>
+const RestaurantView = () => {
+  return (
+    <View>
+      <View style={styles.Map}>
+        <Image source = {require("../../assets/map.png")} style={styles.Image}/>
+      </View>
       <View style={styles.Text}>
         <View style={styles.Location}>
           <Text style={styles.Name}> 
@@ -22,12 +23,27 @@ class Card extends Component {
         </Text>
       </View>
     </View>
-    )
-  }
+  )
 }
 
 const styles = StyleSheet.create({
   Card: {
+    width: 300,
+    height: 300,
+    backgroundColor: "lightblue",
+    display: "flex",
+    borderRadius: 10,
+    overflow: "hidden",
+    shadowColor: "#000",
+    shadowOffset: {
+	    width: 0,
+	    height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 3.84,
+    elevation: 5,
+  },
+  Map: {
     width: 300,
     height: 300,
     backgroundColor: "lightblue",
@@ -70,4 +86,5 @@ const styles = StyleSheet.create({
   }
 })
 
-export default Card;
+export default RestaurantView;
+
